Fixed Issue with RootId being null since the rootID is actually a int64 instead of an int32

+++ Listen360API/IOrganization.cs
@@ -12,7 +12,7 @@ namespace Listen360API
         /// <summary>
         /// Gets or sets the unique identifier of the root organization.
         /// </summary>
-        int RootId { get; }
+        long RootId { get; }
 
         /// <summary>
         /// Gets or sets the unique identifier of the parent organization.
